Version 6 changes:
Move all tag handling into a separate class. This means that all tag
handling can be customized, with the exception of value sharing and
stringrefs. (Because they need access to the internal state of the
decoder)

Suprisingly this will not greatly affect performance of the c-module
since most of the tag handling is reaching back into the cPython
interpreter to create Python objects.

Simple Values are now represented with a python Enum class which fits
their intended use better.