Separate c-module tag handling.
This allows us to turn off built-in tag handling in the C-module so it
is solely done with the python TagHandler class, or any custom class the
user provides.

From testing, a custom Python tag handler called from the c-module is
about %40-%50 of the speed of the built-in tag handlers.

The CBOR2_DISABLE_FAST_TAGS environment variable can be used to control
this.