Fix simple value handling to be more in line with RFC8949
Single byte "Simple Values" (SV for short) less than 20 are just
returned as CBORSimpleValue(n).

SV from 20 to 23 are permitted but when decoding they are returned
as their native python equivalent: True, False, None, and undefined.

SV 24 requires that the following byte is an 8-bit integer between 32
and 255, returned as CBORSimpleValue(n). Any other value is an error.

SV 25,26,27 are prefixes for 2, 4 and 8 bit floats.

SV 28 to 30 are not valid and reserved for future expansion.

SV 31 is the delimiter for indefinite length lists/maps/strings.