Python datetime.isoformat() will output timezones with fractional offset
This is not supported by RFC3339 which requires that timezone offsets be
rounded to the nearest minute.

TODO: Instead of replacing the timestamp, we could modify the time with
the new offset. This would mean the timestamp would still refer to the
same "time". It's a real can of worms.